Considering user feedback
We took feedback from some users wich are not a part of the team, so it was blackbox-testing. The changes are easy to implement but make huge difference such as the right timing the right rounding mor controll structures and so on. In the class Admin I only added a new method everything else is the same.